Complete overhaul of the preferences app
Added Google Device Admin Receiver and integrated a setup wizard for
the device.  Added device management via SMS codes, preferences to
allow changing of the codes, wakeful alarms to do the management work,
and now ICS themed with actionbarsherlock to make it easier to
implement on ics as well as gingerbead.