Added AdminCode= to every SMS Code, Fixed bugs with SIM change, AccessMRS integration
Now all SMS messages need to have the prefix code "AdminCode=".  This
allows for scanning only the initial 10 character substring of the SMS
body to verify if it is admin or not, and not read the entire SMS.
Fixed bugs with SIM changes, allowing for improved performance when
there is no SIM in the phone on boot.  Also fixed AccessMRS integration
in the installation wizard.