Complete rewrite of SIM check and phone lock. Added alarm and warning before phone lock
SIM errors no longer result in silent locking, but present the user
with an alarm, allowing them to turn off the device before the device
locks itself.  SMS is sent to admin phone during this period. SIM
preference is now calculated based on number of missing SIM or number
of changed SIM per given period of time, as set in a number of new
preferences. Preferences UI not done, but the services and defaults are
there.  All of the settings regarding SIM lock are now entirely
customizeable.  Moved this code out of receiver to IntentService. Fixed
potential problem with not reading SIM when airplane mode is on by
toggling off with new receiver. Fixed bug with SMS Service not stopping
SMS delivery for repeating SMS.